# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Maritime Transport System (MTS)",
        "title": "MTS Shipping Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated ports for a given city",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Port"
                ],
                "summary": "List ports by city",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "City ID",
                        "name": "city_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ports list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Port"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing city_id",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/ports/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return details of a specific port",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Port"
                ],
                "summary": "Get port by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Port ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Port found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Port"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Port not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/vessels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated vessels for a specific shipping company",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vessel"
                ],
                "summary": "List vessels by company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shipping company ID",
                        "name": "shipping_company_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vessels list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Vessel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing company ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/vessels/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return details of a specific vessel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vessel"
                ],
                "summary": "Get vessel by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vessel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vessel found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Vessel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Vessel not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/voyages/recommend": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get voyages that can transport required tonnage from start port to end port, sorted by remaining capacity descending",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Voyage"
                ],
                "summary": "Recommend voyages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Start port ID",
                        "name": "start_port_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "End port ID",
                        "name": "end_port_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Required tonnage",
                        "name": "required_ton",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recommended voyages",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service.VoyageRecommendation"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.City": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "city_name": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "delete_time": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "timezone": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "model.Port": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.City"
                        }
                    ]
                },
                "city_id": {
                    "type": "integer"
                },
                "create_time": {
                    "type": "string"
                },
                "delete_time": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "max_draft_meter": {
                    "type": "number"
                },
                "port_code": {
                    "type": "string"
                },
                "port_id": {
                    "type": "integer"
                },
                "port_name": {
                    "type": "string"
                },
                "port_type": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "model.ShippingCompany": {
            "type": "object",
            "properties": {
                "account_status": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "company_id": {
                    "type": "integer"
                },
                "company_name": {
                    "type": "string"
                },
                "contact_person": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "delete_time": {
                    "type": "string"
                },
                "login_password": {
                    "type": "string"
                },
                "login_username": {
                    "type": "string"
                },
                "unified_social_credit_code": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "model.Vessel": {
            "type": "object",
            "properties": {
                "call_sign": {
                    "type": "string"
                },
                "container_teu": {
                    "type": "integer"
                },
                "create_time": {
                    "type": "string"
                },
                "delete_time": {
                    "type": "string"
                },
                "draft_meter": {
                    "type": "number"
                },
                "gross_tonnage": {
                    "type": "number"
                },
                "imo_number": {
                    "type": "string"
                },
                "is_available": {
                    "type": "integer"
                },
                "max_deadweight_ton": {
                    "type": "number"
                },
                "net_tonnage": {
                    "type": "number"
                },
                "shipping_company": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ShippingCompany"
                        }
                    ]
                },
                "shipping_company_id": {
                    "type": "integer"
                },
                "speed_knot": {
                    "type": "number"
                },
                "update_time": {
                    "type": "string"
                },
                "vessel_id": {
                    "type": "integer"
                },
                "vessel_name": {
                    "type": "string"
                },
                "vessel_type": {
                    "type": "string"
                }
            }
        },
...
        "response.PageMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.PageMeta"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "service.VoyageRecommendation": {
            "type": "object",
            "properties": {
                "lineID": {
                    "type": "integer",
                    "format": "int64"
                },
                "lineName": {
                    "type": "string"
                },
                "remainingCapacity": {
                    "type": "number",
                    "format": "float64"
                },
                "vesselID": {
                    "type": "integer",
                    "format": "int64"
                },
                "vesselName": {
                    "type": "string"
                },
                "voyageDate": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}